pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "faiss-cpu>=1.11.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy==1.26.4",
    "openai>=1.88.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "werkzeug>=3.1.3",
]
